refactor: entity => entities utilize makefile to avoid rebuild
